requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.0
greenlet==0.4.17
gunicorn==20.0.4
h11==0.11.0
httptools==0.1.1
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
PyYAML==5.3.1
six==1.15.0
uvicorn==0.13.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
psycopg2-binary
redis
ipython